% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_utils.R
\name{add_err}
\alias{add_err}
\title{Add Error}
\usage{
add_err(
  new_message,
  error_list = error_list,
  location,
  condition = c("skip", "continue")[1]
)
}
\arguments{
\item{new_message}{Error message text}

\item{error_list}{list where errors are stored}

\item{location}{indentify where in the code this error is occuring (i.e what function)}

\item{condition}{character either "skip" or "continue". Skip logs the error
and moves to the next aggreation period. Continue logs the error and attempts to execute the next code block}
}
\description{
Quickly add errors to an error list without all that tedious mucking around with \code{[[]]}
}
\author{
W. S. Drysdale
}
