% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.avg.R
\name{def.avg}
\alias{def.avg}
\title{Define Averaging Period}
\usage{
def.avg(filePaths, fileNames, fileMask, fileDuration, aggregationDuration)
}
\arguments{
\item{filePaths}{vector of file names}

\item{fileNames}{vector of file paths}

\item{fileMask}{string mask that can be passed to \code{as.POSIXct(format = fileMask)}}

\item{fileDuration}{numeric length of files in seconds}

\item{aggregationDuration}{numeric required flux duration}
}
\description{
Determine the files that correspond to the selected flux averaging period using a mask for the file name
}
\author{
W. S. Drysdale
}
