% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.lag.fft.R
\name{def.lag.fft}
\alias{def.lag.fft}
\title{Define lag Time via fft}
\usage{
def.lag.fft(ref, scal, trim = NULL, lagNgtvPstv)
}
\arguments{
\item{ref}{vector of reference data}

\item{scal}{vector of scalar}

\item{trim}{+/- should the corr output be trimmed to in time steps}

\item{lagNgtvPstv}{"n", "p" or "np" - are we searching for negative or positive lags, or should we search in both directions}
}
\description{
Adapted from code from link below. Utilise fft to quickly determine the lag time of a scalar vs reference. \cr
Scalar is expected to be behind the reference data
}
\details{
https://stats.stackexchange.com/questions/147105/r-own-implementation-of-cross-correlation-using-convolution
}
\author{
W. S. Drysdale
}
