% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.lod.R
\name{def.lod}
\alias{def.lod}
\title{Function definition. Flux limit of detection.}
\usage{
def.lod(ref, vars, spcs, rmm, rho_dry, rho_H2O, Lv, freq, conf = 95)
}
\arguments{
\item{ref}{Reference instantaneous data, e.g. w_hor.}

\item{vars}{Data.frame of instantaneous data for LOD calculation.}

\item{spcs}{Chemical species.}

\item{rmm}{Relative molecular masses of chemical species in kg mol-1.}

\item{rho_dry}{Dry air density in mol m-3.}

\item{rho_H2O}{Partial density of H2O.}

\item{Lv}{latent heat of vaporization.}

\item{freq}{Data frequency in Hz.}

\item{conf}{Confidence interval, either 95 or 99\%.}
}
\description{
Function definition. Warpper for Wavelet Eddy-covariance averaging.
}
\author{
Adam R Vaughan \email{adam.vaughan@york.ac.uk}
}
