% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.miss.hndl.R
\name{def.miss.hndl}
\alias{def.miss.hndl}
\title{Define how missing values are handled}
\usage{
def.miss.hndl(eddy.data, para)
}
\arguments{
\item{para}{parameters created by def.uoy.para}
}
\value{
eddy.data with missing values handled
}
\description{
One of:
\itemize{
\item drop rows containing missing values
\item fill missing values with column mean
}
}
\author{
W. S. Drysdale
}
