% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.mton.scal.corr.R
\name{def.mton.scal.corr}
\alias{def.mton.scal.corr}
\title{Definition Function: Motion Scale Correction}
\usage{
def.mton.scal.corr(eddy.data)
}
\arguments{
\item{eddy.data}{eddy.data input data.frame containing veloZaxs, veloZpltf and accZpltf}
}
\description{
Apply motion scale correction to the vertical wind vector (veloZaxs).
Correction is outlined in Prytherch et al. (2015) equation 2.
C. G. Stapleton tested to check there was no difference if applied to raw data vs applied to instantaneous deviations
}
\author{
C. G. Stapleton
}
