% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.para.tmp.R
\name{def.para.tmp}
\alias{def.para.tmp}
\title{Remove scalars from para}
\usage{
def.para.tmp(paraMain, skip_scalar)
}
\arguments{
\item{paraMain}{the main para list}

\item{skip_scalar}{names of species}
}
\description{
Removes the scalars from para that can be skipped this itteration
}
\author{
W. S. Drysdale
}
