% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.pres.h2o.sat.temp.huang.R
\name{def.pres.h2o.sat.temp.huang}
\alias{def.pres.h2o.sat.temp.huang}
\title{Calculate Saturated Water Vapour Pressure (Huang)}
\usage{
def.pres.h2o.sat.temp.huang(temp)
}
\arguments{
\item{temp}{temperature in K}
}
\description{
Calculate Saturated Water Vapour Pressure from Temperature using Huang Formulation
See \href{https://doi.org/10.1175/JAMC-D-17-0334.1}{Huang 2018 JAMC}
}
\author{
W. S. Drysdale
}
