% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{def.rot.mat}
\alias{def.rot.mat}
\title{Define Rotation Matrix}
\usage{
def.rot.mat(degrees)
}
\arguments{
\item{degrees}{rotation in degrees - cw == +ve}
}
\value{
rotation matrix
}
\description{
Defines a rotation matrix in degrees, where a positive value rotates in the clockwise direction
}
\author{
W. S. Drysdale
}
