% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.rtio.mole.h2o.temp.pres.rh.R
\name{def.rtio.mole.h2o.temp.pres.rh}
\alias{def.rtio.mole.h2o.temp.pres.rh}
\title{Calculate dry mole fraction of H2O from temperature, pressure and Relative Humidity}
\usage{
def.rtio.mole.h2o.temp.pres.rh(temp, pres, rh)
}
\arguments{
\item{temp}{air temperature in K}

\item{pres}{air pressure in Pa}

\item{rh}{relative humidity as a percentage (e.g 50 for 50\%)}
}
\description{
Uses the saturation water vapour pressure determined via \link{def.pres.h2o.sat.temp.huang}
along with air temperature (K) and pressure (Pa) to calculate the dry mole fraction
of H2O.
}
\author{
W. S. Drysdale
}
