% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.scales.R
\name{def.scales}
\alias{def.scales}
\title{Definition function: Wrapper function for calculating integral scale lengths and turbulence characteristics.}
\usage{
def.scales(
  REYN,
  lat,
  VarInp = c("veloXaxs", "veloZaxs", "temp", "all")[4],
  sd,
  varScal,
  species = NULL
)
}
\arguments{
\item{REYN}{Output list from eddy4R.turb REYNflux function.}

\item{lat}{Latitude and of class "numeric".}

\item{VarInp}{A vector of class "character" containing the name of variables
to be performed integral turbulence characteristics test. VarInp =
c("veloXaxs","veloZaxs","temp","all"), where "veloXaxs" is along-axis horizontal
wind speed, "veloZaxs" is vertical-axis wind speed, "temp" is air temperature,
and "all" is all three variables. Defaults to "all".}

\item{sd}{A vector or data frame containing standard deviation of VarInp and
of class "numeric". If VarInp = "all", sd shall contain in the follwing orders,
standard deviation of along-axis horizontal wind speed, standard deviation of
vertical-axis wind speed, and standard deviation of air temperature.}

\item{varScal}{A vector or data frame containing the scaling variables of VarInp
and of class "numeric". If VarInp = "all", varScal shall contain in the follwing
orders, scaling variable of wind speed (friction velocity will be used for both
"veloXaxs" and "veloZaxs") and scaling variable of air temperature.}

\item{species}{names of scalars to be passed to \code{def.spcs.name}}
}
\description{
Function defintion. Wrapper function for calculating integral scale lengths and turbulence characteristics.
}
\author{
Adam Vaughan \email{adam.vaughan@york.ac.uk}
}
