% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.valid.input.R
\name{def.valid.input}
\alias{def.valid.input}
\title{Valid Input}
\usage{
def.valid.input(
  eddy.data,
  varsRequired,
  varsCritical,
  species,
  speciesRatioName,
  aggregationPeriod,
  missingThreshold,
  logger
)
}
\arguments{
\item{eddy.data}{input data}

\item{varsRequired}{character vector of columns that must be nominally present to pass def.valid.input()}

\item{varsCritical}{these must have greater than the missingThreshold to pass def.valid.input()}

\item{species}{vector of species names. Match with IntlNatu naming}

\item{speciesRatioName}{created automatically from species. name of gas species following the rtioMoleDry\if{html}{\out{<spc>}} format}

\item{aggregationPeriod}{flux aggregation period (s)}

\item{missingThreshold}{decimal percentage of missing data threshold per file}

\item{logger}{logger object}
}
\description{
Checks input file for several key requirements \cr
\itemize{
\item Number of rows !=0
\item Are required variables present? Skips the file if critcal vals are missing
otherwise flags the optional variables that are missing
\item Are these columns the correct class
\item Are any of the stanard deviations of variables 0? Skip/flag as above
\item Check against the missing data threshold set in para
\item If there is no H20 data, set it to 1e-12.
}
}
