% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.vect.orie.R
\name{def.vect.orie}
\alias{def.vect.orie}
\title{Define Vector Orientation}
\usage{
def.vect.orie(eddy.data, vec_id)
}
\arguments{
\item{eddy.data}{eddy.data data.frame}

\item{vec_id}{character string of format (+/-)(u/v/w)_met}
}
\value{
wind vector column with transformation applied
}
\description{
Apply a transformation to a wind vector based on particular codes: \cr
e.g +veloXaxs will return the positive veloXaxs column, -veloXaxs will return veloXaxs*-1 \cr
Used in wrapper to apply user defined wind vector corrections \cr
Function returns the column only so that the wrapper can handle reasssignment to data.frame \cr
this is to avoid issues surrounding the order in which the function is applied
}
\author{
W. S. Drysdale
}
