% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.wind.dir.flow.R
\name{def.wind.dir.flow}
\alias{def.wind.dir.flow}
\title{Definition function. Create uv_met and d_xy_flow columns}
\usage{
def.wind.dir.flow(eddy.data, freq)
}
\arguments{
\item{eddy.data}{data.frame containing u_met and v_met}

\item{freq}{time resolution of data in Hz}
}
\description{
Used to create/recreate the wind vector dependant uv_met and d_xy_flow columns. \cr
d_xy_flow needs to be recreated for each aggregation period, used during eddy.data subsetting in read.e4r_input() \cr
uv_met and d_xy_flow also need recreating if anemometer definitions are altered. used in wrap.anem.cor()
}
\author{
W. S. Drysdale
}
