% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eddy4R.york-package.R
\docType{package}
\name{eddy4R.york-package}
\alias{eddy4R.york}
\alias{eddy4R.york-package}
\title{eddy4R.york: eddy4R.york}
\description{
Extends the eddy4R suite of packages for use with the WACL-York flux workflows. Depends on the eddy4R suite of packages, and as such is recommended to be used with the associate docker image.
}
\author{
\strong{Maintainer}: Will Drysdale \email{will.drysdale@york.ac.uk}

}
\keyword{internal}
