% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_utils.R
\name{err_skip}
\alias{err_skip}
\title{err_skip}
\usage{
err_skip(error_list)
}
\arguments{
\item{error_list}{error list object}
}
\description{
Does the error_list contain skip conditions
}
\author{
W. S. Drysdale
}
