% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_utils.R
\name{error_list_to_df}
\alias{error_list_to_df}
\title{error list to data.frame}
\usage{
error_list_to_df(error_list, file)
}
\arguments{
\item{error_list}{error list object}

\item{file}{id of the input file where the error occured}
}
\description{
transforms the add_err output to a dataframe that can be more sensibly be written to disk
}
\author{
W. S. Drysdale
}
