% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{log_message}
\alias{log_message}
\title{Log eddy4.york}
\usage{
log_message(
  logger,
  logLevel,
  header,
  periodStartDate,
  periodEndDate,
  error = NULL
)
}
\arguments{
\item{logger}{the logger object}

\item{logLevel}{one of warn, error, info or debug}

\item{header}{character string describing the succinctly describing where in the workflow this is being logged from.}

\item{periodStartDate}{periodStartDate}

\item{periodEndDate}{description}

\item{error}{the simpleError caught by tryCatch}
}
\description{
Uses a logger object created by eddy4R.base::Logger.Singleton and writes a log message of the format:
timestamp log_level: errorType | period: aggregationPeriodText | errorSimple
}
