% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_e4r_input.R
\name{read_e4r_input}
\alias{read_e4r_input}
\title{Read eddy4R input}
\usage{
read_e4r_input(DirInp, agg_f, agg_p, Tz, freq, file_type, PltfEc)
}
\arguments{
\item{DirInp}{input directory}

\item{agg_f}{files to aggregate}

\item{agg_p}{period which files are relvant too}

\item{Tz}{timestamp timezone}

\item{freq}{data aquisiton frequency}

\item{file_type}{either ".csv" or ".rds"}

\item{PltfEc}{platform either "towr" or "airc"}
}
\description{
reads the files relevant for a given flux aggregation period
TODO this can be substantially tidied
}
\author{
W. S. Drysdale
}
