% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input.R
\name{read_input}
\alias{read_input}
\title{Read eddy4R input}
\usage{
read_input(
  DirInp,
  dateFormat,
  filePaths,
  periodStartDate,
  periodEndDate,
  tz,
  freq,
  idepVar,
  PltfEc
)
}
\arguments{
\item{DirInp}{input directory}

\item{dateFormat}{format of date when reading csv for \code{as.POSIXct()} - reading from UNIX time to this format - default \%Y-\%m-\%d \%H:\%M:\%OS"}

\item{filePaths}{files to aggregate}

\item{periodStartDate}{start of aggregation period}

\item{periodEndDate}{end of aggregation period}

\item{tz}{time zone}

\item{freq}{data acquisition frequency}

\item{idepVar}{variable that could be used if Algbase in "trend" or "ord03"
column gets duplicated to a column called "idep" used in \code{eddy4R.turb::def.stat.sta.diff}}

\item{PltfEc}{platform either "towr" or "airc"}
}
\description{
reads the files relevant for a given flux aggregation period
TODO this can be substantially tidied
}
\author{
W. S. Drysdale
}
