% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uoy.towr.ec.R
\name{uoy.towr.ec}
\alias{uoy.towr.ec}
\title{UoY Towr Wrapper, Wrapper}
\usage{
uoy.towr.ec(para, resume = NULL, thshFile = NULL, diagSens = FALSE)
}
\arguments{
\item{thshFile}{The file directory where the threshold table are being saved. Default as NULL.}

\item{diagSens}{Logical to state if the sensor diqgnostic flags are calculated. Default as FALSE.}
}
\description{
Handles file reading, progress and error logging. Calls wrap.uoy.ec.towr on valid input files.
Can resume the loop from an agg_files index, supplied to resume
}
\author{
W. S. Drysdale
}
