% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.anem.cor.R
\name{wrap.anem.cor}
\alias{wrap.anem.cor}
\title{Anemometer Corrections Wrapper}
\usage{
wrap.anem.cor(eddy.data, para)
}
\arguments{
\item{eddy.data}{eddy.data data.frame}

\item{para}{parameters list}
}
\description{
Apply transformations to wind vectors easily. Steps in order: \cr
Use def.vect.orie to reorient and invert vectors - para$u_met,v_met,w_met \cr
Use def.rot.mat to rotate around the uv plane - para$anemometer_offset \cr
Apply W-boost correction to Gill Anemometers - para$w_boost
}
\author{
W. S. Drysdale
}
