% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.foot.k04.R
\name{wrap.foot.k04}
\alias{wrap.foot.k04}
\title{Wrap K04 Footprint model}
\usage{
wrap.foot.k04(
  angZaxsErth,
  footprintResolutionM,
  veloYaxsHorSd,
  veloZaxsHorSd,
  veloFric,
  distZaxsMeasDisp,
  distZaxsAbl,
  distZaxsRgh,
  footprintCumulativeThreshold,
  univFunc
)
}
\arguments{
\item{angZaxsErth}{wind direction to rotate the inertial footprint matrix}

\item{footprintResolutionM}{Resolution of footprint model in meters (default 100)}

\item{veloYaxsHorSd}{crosswind fluctuations}

\item{veloZaxsHorSd}{vertical wind fluctuations}

\item{veloFric}{friction velocity}

\item{distZaxsMeasDisp}{height of measurement - displacement}

\item{distZaxsAbl}{boundary layer height}

\item{distZaxsRgh}{Roughness length}

\item{footprintCumulativeThreshold}{threshold for cumulative footprint extent (default 0.8)}

\item{univFunc}{integral over the stability-dependent universal function to make the log-wind-profile applicable to different atmospheric stratifications; from def.func.univ()}
}
\description{
Wraps the K04 (\code{eddy4R.turb::def.foot.k04}) footprint model to pass
params and tidy output into data.frame. Outputs data.frame with x, y and value columns
x and y are in m (based off of footprintResolutionM).
}
\author{
W. S. Drysdale
}
