% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.isca.R
\name{wrap.isca}
\alias{wrap.isca}
\title{Wrapper Integral Length Scales}
\usage{
wrap.isca(REYN, species, speciesRatioName, PltfEc)
}
\arguments{
\item{REYN}{requires REYN$data and REYN$diff}

\item{species}{species names}

\item{speciesRatioName}{species mixing ration column names}

\item{PltfEc}{platform - defined scalEddy for \code{def.dist.isca()} - currently only supports "towr"}
}
\description{
Wraps \code{def.dist.isca()}
}
