% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.lag.R
\name{wrap.lag}
\alias{wrap.lag}
\title{UoY Lag definition wrapper}
\usage{
wrap.lag(
  eddy.data,
  lagVars,
  lagApplyRangeLimit,
  lagRangeLimit,
  lagDefaults,
  lagNOc,
  freq,
  freqThsh = NULL,
  speciesRatioName,
  lagMax,
  lagNgtvPstv
)
}
\arguments{
\item{eddy.data}{eddy.data object}

\item{lagVars}{created from species plus temperature and water vapour}

\item{lagApplyRangeLimit}{Should a window be applied to the ccf lag?}

\item{lagRangeLimit}{vector or list of vectors containing boundary for acceptable lags}

\item{lagDefaults}{numeric value or vector to apply to lag if ccf falls outside of range or is not used. If this is NA, then the lag will be the value of maxmimum absolute ACF within the given lag ranges}

\item{lagNOc}{when processing NOx, is NO2 actually NOc and therfore CE applied after lagging. T/F}

\item{freq}{data aquisition frequency of input data}

\item{freqThsh}{cutoff frequency for the low end of the high pass filter. defaults to 1 / (2 * lagMax / freq)}

\item{speciesRatioName}{created automatically from species. name of gas species following the rtioMoleDry\if{html}{\out{<spc>}} format}

\item{lagMax}{maximum lag range in timesteps (Hz) to search for a lag over. Defaults to 40*freq}

\item{lagNgtvPstv}{what "direction" can lag occur? "n","p","np"}
}
\description{
Wrapper for def.lag to use in UoY workflows. Returns data lag times and acf information
}
