% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.mton.scale.corr.R
\name{wrap.mton.scale.corr}
\alias{wrap.mton.scale.corr}
\title{Motion Scale Correction Wrapper}
\usage{
wrap.mton.scale.corr(eddy.data)
}
\arguments{
\item{eddy.data}{eddy.data data.frame}
}
\description{
Apply motion scale correction to the vertical wind vector (veloZaxs).
Correction is outlined in Prytherch et al. (2015) equation 2.
C. G. Stapleton tested to check there was no difference if applied to raw data vs applied to instantaneous deviations
}
\author{
C. G. Stapleton
}
