% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.voc.qf.R
\name{wrap.qf.voc}
\alias{wrap.qf.voc}
\title{Wrapper function: Determination of statistical flags for VOC flux comparison}
\usage{
wrap.qf.voc(data, thshFile = NULL)
}
\arguments{
\item{data}{Required input. A data frame containing the data to be evaluated.}

\item{thshFile}{The file directory where the threshold table are being saved.}
}
\value{
qfPlau A list of data frame containing the calculated sensors quality flags and statistical flags
}
\description{
Wrapper function to generate the statistical flags for sensors (PTR-Qi-TOF and CPEC200) which used in VOC comparison.
}
\references{
License: GNU AFFERO GENERAL PUBLIC LICENSE Version 3, 19 November 2007.
}
\seealso{
Currently none
}
\author{
Natchaya Pingintha-Durden \email{ndurden@battelleecology.org}
}
\keyword{VOC,}
\keyword{qfqm}
