% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.rot.R
\name{wrap.rot}
\alias{wrap.rot}
\title{Define Rotation of Wind Vectors}
\usage{
wrap.rot(
  data,
  MethRot = c("single", "double", "plnrFit", "autoPlnrFit", "none")[1],
  plnrFitCoef = NULL,
  plnrFitType = c("simple", "time", "wind")[1]
)
}
\arguments{
\item{data}{data.frame containing u_met, v_met and w_met \cr
vectors should be defined correspondingly as u/v/w == E/N/U == x/y/z \link{data.frame}}

\item{MethRot}{method of rotation to be used, one of: \itemize{
\item "single" - rotate into the mean wind
\item "double" - apply single rotation and additionally rotate to align w and z (minimise w)
\item "planarFit" - apply the coefficeints for planat fit, supplied to plnrFitCoef. \cr
Note that these coefficients should be determined in the coordinate system they are applied to.
\item "autoPlanrFit" - determine and apply the planar fit coefficeints from data.
\item "none" - perform no rotation
}}

\item{plnrFitCoef}{coefficients for planar fit. \link{numeric vector or data.frame} \cr
Depending on plnrFitType, plnrFitCoef should be supplied as the following structure: \itemize{
\item simple - numeric vector constant of coefficeients, or coefficeients
that are controlled from the workflow. c(al,be,b0)
\item time - data.frame with columns date, al, be, b0. values with date nearest to mean(data$date) will be used
\item wind - data.frame with columns PSI_uv, al, be, b0. values with PSI_uv nearest to average PSI_uv will be used
}}

\item{plnrFitType}{type of planar fit, "simple", "date" or "wind". \link{character vector}}
}
\value{
Data object with rotated wind vectors names as u/v/w_rot
}
\description{
Perform various types of wind vector rotations incl. single, double, planar fit.
}
\references{
Code adapted from REYNFlux_P5 (35ceda9) and flow.turb.tow.neon.dp04.r (914a9e9) \cr
Stefan Metzger / Dave Durden / Natchaya P-Durden / Cove Sturtevant / Ke Xu
}
\author{
W. S. Drysdale
}
