% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.towr.R
\name{wrap.towr}
\alias{wrap.towr}
\title{Tower Wrapper}
\usage{
wrap.towr(paraMain, resume = NULL, thshFile = NULL, diagSens = FALSE)
}
\arguments{
\item{paraMain}{parameters list made by \code{def.para()}}

\item{resume}{loop iterator to restart at.}

\item{thshFile}{The file directory where the threshold table are being saved. Default as NULL.}

\item{diagSens}{Logical to state if the sensor diqgnostic flags are calculated. Default as FALSE.}
}
\description{
Handles file reading, progress and error logging.
}
\author{
W. S. Drysdale
}
