% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.uoy.despike.R
\name{wrap.uoy.despike}
\alias{wrap.uoy.despike}
\title{Wrap UoY Despiking}
\usage{
wrap.uoy.despike(eddy.data, despike_vars, despike_threshold, verbose)
}
\arguments{
\item{verbose}{}
}
\description{
Contains the despiking loop
}
\author{
W. S. Drysdale
}
