% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.REYN.R
\name{write.REYN.csv}
\alias{write.REYN.csv}
\title{Write REYN csv}
\usage{
write.REYN.csv(REYN, lag_time, ACF, para, file_name, qfOut = NULL)
}
\arguments{
\item{REYN}{REYN List of objects at the end of calcualtions}

\item{lag_time}{lag time data frame}

\item{ACF}{ACF data.frame}

\item{para}{parameters list}

\item{file_name}{file name based on aggreation period}

\item{qfOut}{A list of data frame containing the calculated sensors quality flags and statistical flags. Defaults as NULL.}
}
\description{
Writes REYN output as RDS objects - as Will's workflow
}
