% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assign_input_units}
\alias{assign_input_units}
\title{Assign Unit Inputs}
\usage{
assign_input_units(eddy.data, para)
}
\arguments{
\item{eddy.data}{input data}

\item{para}{parameter list}
}
\description{
Takes the input list from \code{def.para()} or \code{default_unit_list()}
and sets the corresponding columns in eddy.data.
}
