% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.lod.from.wrap.lag.R
\name{def.lod.from.wrap.lag}
\alias{def.lod.from.wrap.lag}
\title{Limit of Detection from wrap.lag output}
\usage{
def.lod.from.wrap.lag(lag_out, REYN, freq)
}
\arguments{
\item{lag_out}{output of wrap.lag}

\item{REYN}{eddy4R REYN list - requires mean and diff}

\item{freq}{measurement frequency}
}
\description{
Uses the output of \code{wrap.lag} to calculate the LOD followling Langford et al 2015 and \code{def.lod}.
}
