% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.lod.R
\name{def.lod}
\alias{def.lod}
\title{Flux Limit of Detection}
\usage{
def.lod(REYN, measCol, freq, refCol = "veloZaxs")
}
\arguments{
\item{REYN}{eddy4R REYN list - requires mean and diff}

\item{measCol}{columns to calculate LOD for}

\item{freq}{measurement frequency}

\item{refCol}{column to lag against, default veloZaxs}
}
\description{
Langford et al., 2015 Eddy-covariance data with low signal-to-noise ratio method
of flux Limit of detection. Based on original implementation by Adam Vaughan
}
