% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.miss.hndl.R
\name{def.miss.hndl}
\alias{def.miss.hndl}
\title{Define how missing values are handled}
\usage{
def.miss.hndl(
  eddy.data,
  missingMethod,
  missingThreshold,
  aggregationPeriod,
  freq
)
}
\arguments{
\item{eddy.data}{eddy.data data frame}

\item{missingMethod}{how should missing data be handeled if it is less than the threshold. "drop","mean"}

\item{missingThreshold}{decimal percentage of missing data threshold per file}

\item{aggregationPeriod}{flux aggregation period (s)}

\item{freq}{data aquisition frequency of input data}
}
\value{
eddy.data with missing values handled
}
\description{
One of:
\itemize{
\item drop rows containing missing values
\item fill missing values with column mean
}
}
\author{
W. S. Drysdale
}
