% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.para.R
\name{def.para}
\alias{def.para}
\title{Define eddy4R.york Parameters}
\usage{
def.para(
  file_duration = 3600,
  file_mask = "NOx_5Hz_yymmdd_HHMM0_170322_000015_cor_temp.nc",
  species,
  freqIN = 5,
  freqOUT = 5,
  files = NULL,
  Tz = "GMT",
  file_type_in = ".csv",
  required_para = c("date", "t_utc", "d_z_m", "d_xy_flow", "p_air", "d_z_ABL"),
  critical_variable = c("u_met", "v_met", "w_met", "T_air", "uv_met"),
  SND_correct = FALSE,
  AlgBase = "trnd",
  agg_period = 3600,
  missing_thresh = 0.1,
  missing_method = c("drop", "mean")[1],
  noc_lag = F,
  tempHead = c("T_sonic", "T_air", "T_v_0")[2:3],
  anem_type = c("other", "CSAT3"),
  anemometer_offset = NULL,
  u_met = "+u_met",
  v_met = "+v_met",
  w_met = "+w_met",
  w_boost = F,
  lag_correction = T,
  determine_lag = T,
  lag_type = c("ccf", "fft")[1],
  lagNgtvPstv = c("n", "p", "np")[1],
  absolute_lag = -10,
  restrict_lag_range = F,
  lag_boundary = NULL,
  despike = T,
  despike_threshold = c(5, 8, 8),
  MethRot = c("single", "double", "planarFit")[2],
  plnrFitCoef = NULL,
  plnrFitType = c("simple", "time", "wind")[1],
  wlp_cor = F,
  plot_acf = F,
  highfreq_cor = T,
  file_id = "NOx_5Hz",
  run_id = "BT_Tower",
  file_type_out = c("csv"),
  site_name = NULL,
  analysis,
  write_fast_data = F,
  DirInp,
  DirWrk,
  DirOut = NULL,
  DirFast = NULL,
  PltfEc = "towr",
  TimeDiffUtcLt = 0,
  ZoneUtm = data.frame(Zone = 30, Estg = 698478, Nthg = 5711690),
  ElevAslTow = 35,
  ElevAglSens = 177,
  ElevAglDisp = 0,
  Lat = 51.521381,
  first_file_begin = NULL,
  final_file_begin = NULL,
  cross_correlation_vars = NULL,
  despike_vars = NULL,
  wavelet_ec = FALSE,
  wavelet_av = "mean",
  wavelet_win = 5,
  wavelet_agg = 1,
  wavelet_cores = 2,
  wavelet_maxscale = 60,
  footprint = F,
  ...
)
}
\arguments{
\item{file_duration}{expected duration of a complete input file \link{Hz}}

\item{file_mask}{mask to pass mask_extract_date for detection of start date}

\item{species}{vector of species names. Match with IntlNatu naming to auto complete species_RMM}

\item{freqIN}{input frequency \link{Hz}}

\item{freqOUT}{resample frequency \link{Hz}}

\item{files}{list of file names}

\item{Tz}{time zone Olsen name}

\item{file_type_in}{input file type, currently only supports .csv}

\item{required_para}{character vector of columns that must be nominally present to pass def.valid.input()}

\item{critical_variable}{these must have greater than the missing_thresh to pass def.valid.input()}

\item{AlgBase}{detrending method for def.base.ec "mean","trend","ord3"}

\item{agg_period}{flux aggregation period \link{s}}

\item{missing_thresh}{decimal percentage of missing data threshold per file}

\item{missing_method}{how should missing data be handeled if it is less than the threshold. "drop","mean"}

\item{noc_lag}{when processing NOx, is NO2 actually NOc and therfore CE applied after lagging. T/F}

\item{tempHead}{names of headers containg temperature data. default c("T_air","T_v_0"). \link{character vector}}

\item{anem_type}{"other" or "CSAT3". In the case of CSAT 3, use the def.met.body() function. Otherwise use def.vect.orie(). \cr
For CSAT 3, anemometer offset should be positve from north. For other, anemomet offset should be +ve for clockwise correction, \cr
negative for anticlockwise}

\item{anemometer_offset}{rotation angle for uv plane of wind vectors \link{degrees}}

\item{u_met}{wind vector definition for wrap.anem.cor}

\item{v_met}{wind vector definition for wrap.anem.cor}

\item{w_met}{wind vector definition for wrap.anem.cor}

\item{lag_correction}{Should lag correction be undertaken? T/F}

\item{determine_lag}{Should the lag be determined via crosscorrelation? T/F}

\item{lag_type}{Calculate lag via def.lag or def.laf.fft. "ccf"/"fft"}

\item{lagNgtvPstv}{what "direction" can lag occur? "n","p","np"}

\item{absolute_lag}{numeric value or vector to apply to lag if ccf falls outside of range or is not used}

\item{restrict_lag_range}{Should a window be applied to the ccf lag?}

\item{lag_boundary}{vector or list of vectors containing boundary for acceptable lags}

\item{despike}{Should despiking be undertaken? T/F}

\item{despike_threshold}{numeric vector of standard deviations for temperature and water vapour, plus all species}

\item{MethRot}{type of coordinate rotation of wind vectors to be performed. "single","double","planarFit"}

\item{plnrFitCoef}{coefficients for planar fit calculation. \cr
either numeric vector for plnrFitType = "simple" or data.frame for plnrFitType = "wind" or "time"}

\item{plnrFitType}{type of planar fit, "simple", "date" or "wind". (character vector) \itemize{
\item simple - numeric vector constant of coefficeients, or coefficeients that are controlled from the workflow. c(al,be,b0)
\item time - data.frame with columns date, al, be, b0. values with date nearest to mn$date are used
\item time - data.frame with columns PSI_uv, al, be, b0. values with date nearest to mn$PSI_uv are used}}

\item{wlp_cor}{Should fast water vapour be used to correct NO and NO2 concentrations.}

\item{plot_acf}{Plot the lag output for each aggregation period for each species}

\item{highfreq_cor}{Should high frequency corrections be performed}

\item{file_id}{Used in file nameing}

\item{run_id}{Used in file nameing}

\item{file_type_out}{output file type, currently only supports .csv}

\item{site_name}{Used in file nameing}

\item{analysis}{Used in file nameing}

\item{write_fast_data}{should high frequency intermediates be writen (imfl,data,base outputs from REYN)}

\item{DirWrk}{root of the data directory}

\item{DirOut}{by default is created as \code{file.path(DirWrk,"out",site_name, run_id, analysis)}, can be overridden here.}

\item{DirFast}{directory where to save fast data outputs. Usuallt just within DirOut but can be overridden here.}

\item{PltfEc}{switch for eddy4r functions between tower and aircraft. towr,airc}

\item{TimeDiffUtcLt}{offset of timezone from UTC}

\item{ZoneUtm}{data.frame containing UTM info e.g data.frame(Zone=30, Estg=698478, Nthg=5711690)}

\item{ElevAslTow}{Tower elevation above sealevel}

\item{ElevAglSens}{Tower elevation above ground}

\item{ElevAglDisp}{Displacement Height}

\item{Lat}{latitude of tower location}

\item{first_file_begin}{used to aid def.avg - should be created internally}

\item{final_file_begin}{used to aid def.avg - should be created internally}

\item{cross_correlation_vars}{created from species plus temperature and water vapour}

\item{despike_vars}{created from species plus temperature and water vapour}

\item{wavelet_ec}{supply argument true or false for conducting wavelet eddy-covariance}

\item{wavelet_av}{supply argument "mean" or "band" for conducting wavelet eddy-covariance averaging.}

\item{wavelet_win}{supply value for data window to use for wavelet averaging. Tower setup give value in minutes, e.g. 5. Aircraft give distance in meters.}

\item{wavelet_agg}{supply value for data jump to use for wavelet averaging. Tower setup give value in minutes, e.g. 1. Aircraft gives distance in meters.}

\item{wavelet_cores}{specify number of CPU cores to use for parallel wavelet averaging.}

\item{wavelet_maxscale}{maximum wavlet scale to use for averaging. Default is 60 minutes. Give value in in minutes, e.g. 60.}

\item{footprint}{Footprint modelling.}

\item{...}{supply arguments that have not been given an explicit definition yet here}

\item{docker_name}{docker username so that file paths can be found/created}

\item{DirIn}{must be supplied - relative to DirWrk}
}
\description{
Create para list, use arguments to overide defaults
}
\author{
W. S. Drysdale
}
