% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.para.R
\name{def.para}
\alias{def.para}
\title{Define eddy4R.york Parameters}
\usage{
def.para(
  DirInp,
  DirWrk,
  runID,
  analysis,
  fileMask,
  fileDuration,
  DirOut = NULL,
  siteName = NULL,
  subDir = c("none", "monthly", "daily")[1],
  species = NULL,
  speciesRatioName = NULL,
  speciesFluxName = NULL,
  unitList = NULL,
  freq = 5,
  files = NULL,
  filePattern = ".csv",
  tz = "UTC",
  dateFormat = "\%Y-\%m-\%d \%H:\%M:\%OS",
  varsRequired = c("unixTime", "distZaxsMeas", "presAtm", "distZaxsAbl"),
  varsCritical = c("veloXaxs", "veloYaxs", "veloZaxs", "tempAir"),
  AlgBase = "trnd",
  idepVar = "unixTime",
  aggregationDuration = 3600,
  missingThreshold = 0.1,
  missingMethod = c("drop", "mean")[1],
  lagNOc = F,
  anemometerType = c("other", "CSAT3"),
  anemometerOffset = NULL,
  veloXaxs = "+veloXaxs",
  veloYaxs = "+veloYaxs",
  veloZaxs = "+veloZaxs",
  wBoost = F,
  lagApplyCorrection = T,
  lagNgtvPstv = c("n", "p", "np")[1],
  lagDefaults = -10,
  lagApplyRangeLimit = F,
  lagRangeLimit = NULL,
  lagVars = NULL,
  despike = T,
  despikeThreshold = c(5, 8, 8),
  despikeVars = NULL,
  MethRot = c("single", "double", "planarFit")[2],
  plnrFitCoef = NULL,
  plnrFitType = c("simple", "time", "wind")[1],
  lat,
  PltfEc = "towr",
  ZoneUtm = data.frame(Zone = 30, Estg = 698478, Nthg = 5711690),
  fileFirstStart = NULL,
  fileLastStart = NULL,
  spectralTaperingWeight = 0.5,
  footprintResolutionM = 100,
  footprintCumulativeThreshold = 0.8,
  wavelet_ec = FALSE,
  wavelet_av = "mean",
  wavelet_win = 5,
  wavelet_agg = 1,
  wavelet_cores = 2,
  wavelet_maxscale = 60,
  ...
)
}
\arguments{
\item{DirInp}{must be supplied - relative to DirWrk}

\item{DirWrk}{root of the data directory}

\item{runID}{Used in file nameing}

\item{analysis}{Used in file nameing}

\item{fileMask}{string mask that can be passed to \code{as.POSIXct(format = mask)} to decode the date in the file name}

\item{fileDuration}{expected duration of a complete input file (Hz)}

\item{DirOut}{by default is created as \code{file.path(DirWrk,"out",siteName, runID, analysis)}, can be overridden here.}

\item{siteName}{Used in file nameing}

\item{subDir}{one of c("none", "monthly", "daily") - default none. Should the outputs be split into monthly or daily subdirectories}

\item{species}{vector of species names. Match with IntlNatu naming}

\item{speciesRatioName}{created automatically from species. name of gas species following the rtioMoleDry\if{html}{\out{<spc>}} format}

\item{speciesFluxName}{created automatically from species. name of gas species following the flux\if{html}{\out{<spc>}} format}

\item{unitList}{defaults to \code{default_unit_list()} plus species field based off of species argument
speices is a list with names that match speciesRatioName describing the units per species. Defaults to mol\if{html}{\out{<spc>}} mol-1Dry}

\item{freq}{data aquisition frequency of input data}

\item{files}{list of file names}

\item{filePattern}{pattern to help filter input directory - default .csv}

\item{tz}{default "UTC".}

\item{dateFormat}{default \%Y-\%m-\%d \%H:\%M:\%OS"}

\item{varsRequired}{character vector of columns that must be nominally present to pass def.valid.input()}

\item{varsCritical}{these must have greater than the missingThreshold to pass def.valid.input()}

\item{AlgBase}{detrending method for def.base.ec "mean","trend","ord03"}

\item{idepVar}{required if Algbase is trend or ord03. Name of column
containing independant variable for determining the base function. Column
gets duplicated to a column called "idep" use use in
\code{eddy4R.turb::def.stat.sta.diff}. Default unixTime.}

\item{aggregationDuration}{flux aggregation period (s)}

\item{missingThreshold}{decimal percentage of missing data threshold per file}

\item{missingMethod}{how should missing data be handeled if it is less than the threshold. "drop","mean"}

\item{lagNOc}{when processing NOx, is NO2 actually NOc and therfore CE applied after lagging. T/F}

\item{anemometerType}{"other" or "CSAT3". In the case of CSAT 3, use the def.met.body() function. Otherwise use def.vect.orie(). \cr
For CSAT 3, anemometer offset should be positve from north. For other, anemomet offset should be +ve for clockwise correction, \cr
negative for anticlockwise}

\item{anemometerOffset}{rotation angle for uv plane of wind vectors (degrees)}

\item{veloXaxs}{wind vector definition for wrap.anem.cor}

\item{veloYaxs}{wind vector definition for wrap.anem.cor}

\item{veloZaxs}{wind vector definition for wrap.anem.cor}

\item{wBoost}{should w boost correction for certain Gill Anemometers be applied? Default false}

\item{lagApplyCorrection}{Should lag correction be undertaken? T/F}

\item{lagNgtvPstv}{what "direction" can lag occur? "n","p","np"}

\item{lagDefaults}{numeric value or vector to apply to lag if ccf falls outside of range or is not used}

\item{lagApplyRangeLimit}{Should a window be applied to the ccf lag?}

\item{lagRangeLimit}{vector or list of vectors containing boundary for acceptable lags}

\item{lagVars}{created from species plus temperature and water vapour}

\item{despike}{Should despiking be undertaken? T/F}

\item{despikeThreshold}{numeric vector of standard deviations for temperature and water vapour, plus all species}

\item{despikeVars}{created from species plus temperature and water vapour}

\item{MethRot}{type of coordinate rotation of wind vectors to be performed. "single","double","planarFit"}

\item{plnrFitCoef}{coefficients for planar fit calculation. \cr
either numeric vector for plnrFitType = "simple" or data.frame for plnrFitType = "wind" or "time"}

\item{plnrFitType}{type of planar fit, "simple", "date" or "wind". (character vector) \itemize{
\item simple - numeric vector constant of coefficeients, or coefficeients that are controlled from the workflow. c(al,be,b0)
\item time - data.frame with columns date, al, be, b0. values with date nearest to mn$date are used
\item time - data.frame with columns PSI_uv, al, be, b0. values with date nearest to mn$PSI_uv are used}}

\item{lat}{latitude of tower location}

\item{PltfEc}{switch for eddy4r functions between tower and aircraft. towr,airc}

\item{ZoneUtm}{data.frame containing UTM info e.g data.frame(Zone=30, Estg=698478, Nthg=5711690)}

\item{fileFirstStart}{used to aid def.avg - should be created internally}

\item{fileLastStart}{used to aid def.avg - should be created internally}

\item{spectralTaperingWeight}{tapering value to be supplied to \code{eddy4R.turb::def.spec.fft.fwd()}
and ultimately \code{stats::spec.taper()}. value < 0.5 or FALSE
Default 0.05 as per \code{eddy4R.turb::def.spec.fft.fwd()}.}

\item{footprintResolutionM}{Resolution of footprint model in meters (default 100)}

\item{footprintCumulativeThreshold}{threshold for cumulative footprint extent (default 0.8)}

\item{wavelet_ec}{supply argument true or false for conducting wavelet eddy-covariance}

\item{wavelet_av}{supply argument "mean" or "band" for conducting wavelet eddy-covariance averaging.}

\item{wavelet_win}{supply value for data window to use for wavelet averaging. Tower setup give value in minutes, e.g. 5. Aircraft give distance in meters.}

\item{wavelet_agg}{supply value for data jump to use for wavelet averaging. Tower setup give value in minutes, e.g. 1. Aircraft gives distance in meters.}

\item{wavelet_cores}{specify number of CPU cores to use for parallel wavelet averaging.}

\item{wavelet_maxscale}{maximum wavlet scale to use for averaging. Default is 60 minutes. Give value in in minutes, e.g. 60.}

\item{...}{supply arguments that have not been given an explicit definition yet here}
}
\description{
Create para list, use arguments to overide defaults
}
\author{
W. S. Drysdale
}
