% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.scales.R
\name{def.scales}
\alias{def.scales}
\title{Definition function: Wrapper function for calculating integral scale lengths and turbulence characteristics.}
\usage{
def.scales(
  REYN,
  lat,
  VarInp = c("veloXaxs", "veloZaxs", "temp", "all")[4],
  sd,
  varScal,
  species = NULL
)
}
\arguments{
\item{\code{REYN}}{Output list from eddy4R.turb REYNflux function. \link{-}}

\item{\code{WAVE}}{Output list from eddy4R.turb Wavelet analysis. \link{-}}

\item{\code{VarInp}}{A vector of class "character" containing the name of variables to be performed integral turbulence characteristics test. VarInp = c("veloXaxs","veloZaxs","temp","all"), where "veloXaxs" is along-axis horizontal wind speed, "veloZaxs" is vertical-axis wind speed, "temp" is air temperature, and "all" is all three variables. Defaults to "all".\link{-}

Changelog
Will Drysdale (2019-01-29)
Removed SiteInfo dependance, added switches as explicit variables
Natchaya Pingintha-Durden (2019-09-09)
added determination of qfItc for each flux
Natchaya Pingintha-Durden (2019-10-24)
update to use itc from w_hor to represent itc of F_H_en}
}
\description{
Function defintion. Wrapper function for calculating integral scale lengths and turbulence characteristics.
}
\author{
Adam Vaughan \email{adam.vaughan@york.ac.uk}
}
