% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/def.valid.input.R
\name{def.valid.input}
\alias{def.valid.input}
\title{Valid Input}
\usage{
def.valid.input(eddy.data, para, aggregationPeriod, logger)
}
\arguments{
\item{eddy.data}{input data}

\item{para}{params list}

\item{aggregationPeriod}{\code{aggregationPeriod[i,]}}

\item{logger}{logger object}
}
\description{
Checks input file for several key requirements \cr
\itemize{
\item Number of rows !=0
\item Are required variables present? Skips the file if critcal vals are missing
otherwise flags the optional variables that are missing
\item Are these columns the correct class
\item Are any of the stanard deviations of variables 0? Skip/flag as above
\item Check against the missing data threshold set in para
\item If there is no H20 data, set it to 1e-12.
}
}
