% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{default_unit_list}
\alias{default_unit_list}
\title{Default Unit List}
\usage{
default_unit_list()
}
\description{
Returns a list containing the default units for temperature pressure and
velocity. Useful as a base if the user needs to change the defaults in
\code{def.para()$unitList}
}
