% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_input.R
\name{read_input}
\alias{read_input}
\title{Read eddy4R input}
\usage{
read_input(DirInp, dateFormat, agg_f, agg_p, Tz, freq, idepVar, PltfEc)
}
\arguments{
\item{DirInp}{input directory}

\item{dateFormat}{format of date when reading csv for \code{as.POSIXct()}}

\item{agg_f}{files to aggregate}

\item{agg_p}{period which files are relvant too}

\item{Tz}{timestamp timezone}

\item{freq}{data aquisiton frequency}

\item{idepVar}{variable that could be used if Algbase in "trend" or "ord03"
column gets duplicated to a column called "idep" use use in \code{eddy4R.turb::def.stat.sta.diff}}

\item{PltfEc}{platform either "towr" or "airc"}
}
\description{
reads the files relevant for a given flux aggregation period
TODO this can be substantially tidied
}
\author{
W. S. Drysdale
}
