% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.anem.cor.R
\name{wrap.anem.cor}
\alias{wrap.anem.cor}
\title{Anemometer Corrections Wrapper}
\usage{
wrap.anem.cor(eddy.data, anemometerOffset, wBoost)
}
\arguments{
\item{eddy.data}{eddy.data data.frame}

\item{anemometerOffset}{rotation angle for uv plane of wind vectors (degrees)}

\item{wBoost}{should w boost correction for certain Gill Anemometers be applied? Default false}
}
\description{
Apply transformations to wind vectors easily. Steps in order: \cr
Use def.rot.mat to rotate around the uv plane - anemometerOffset \cr
Apply W-boost correction to Gill Anemometers - wBoost
}
\author{
W. S. Drysdale
}
