% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.despike.R
\name{wrap.despike}
\alias{wrap.despike}
\title{Wrap UoY Despiking}
\usage{
wrap.despike(eddy.data, despike_vars, despike_threshold, verbose)
}
\arguments{
\item{eddy.data}{eddy.data object}

\item{despike_vars}{character vector of variables to be despiked}

\item{despike_threshold}{number of standard deviations to qualify as a spike}

\item{verbose}{boolean - when false the messages from \code{def.dspk.br86} are suppressed}
}
\description{
Contains the despiking loop
}
\author{
W. S. Drysdale
}
