% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.lag.R
\name{wrap.lag}
\alias{wrap.lag}
\title{UoY Lag definition wrapper}
\usage{
wrap.lag(eddy.data, para, agg_count)
}
\arguments{
\item{eddy.data}{eddy.data object}

\item{para}{parameter list from \code{def.uoy.para}}

\item{agg_count}{to use when implementing drifting lags - to do}
}
\description{
Wrapper for def.lag to use in UoY workflows. Returns data lag times and acf information
}
