% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.spec.R
\name{wrap.spec}
\alias{wrap.spec}
\title{Wrap Spectral Analysis}
\usage{
wrap.spec(eddy.data, speciesRatioName, freq, spectralTaperingWeight = 0.05)
}
\arguments{
\item{eddy.data}{time domain data including "veloXaxsInp", "veloYaxsInp", "veloZaxs" and "tempAir".
It is expected that lag corrections and other QA/QC has been
performed on these data already.}

\item{speciesRatioName}{chemical species for flux calculations, can be NULL if
no species are present.}

\item{freq}{data aquisition frequency of input data}

\item{spectralTaperingWeight}{tapering value to be supplied to \code{eddy4R.turb::def.spec.fft.fwd()}
and ultimately \code{stats::spec.taper()}. value < 0.5 or FALSE
Default 0.05 as per \code{eddy4R.turb::def.spec.fft.fwd()}.}
}
\description{
Computes the power and co-spectra using \code{eddy4R.turb::def.spec.fft.fwd()}
and \code{eddy4R.turb::def.spec.cosp()}. Output is then tidied into data.frame
containing measured, relative (0-1) and normalised (f*(U/(z-d))) frequency,
as well as the power and cospectra, which are marked by the "spectrumType" column
}
\author{
W. S. Drysdale
}
