% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrap.towr.R
\name{wrap.towr}
\alias{wrap.towr}
\title{Wrap Tower}
\usage{
wrap.towr(
  eddy.data,
  para,
  file_count,
  skip_scalar,
  verbose = FALSE,
  progress_bar,
  agg_period,
  thshFile = NULL,
  diagSens = FALSE
)
}
\arguments{
\item{eddy.data}{input data}

\item{para}{params list}

\item{file_count}{file number (used for progress only)}

\item{skip_scalar}{character vector of any scalars to be skipped for this file only}

\item{verbose}{boolean, suppresses some chatty functions}

\item{progress_bar}{progress bar object}

\item{thshFile}{The file directory where the threshold table are being saved. Default as NULL.}

\item{diagSens}{Logical to state if the sensor diqgnostic flags are calculated. Default as FALSE.}
}
\description{
University of York Tower Eddy-Covariance turbulence processing.
Based on earlier workflows with contributions from: \cr
Adam R Vaughan \cr
Stefan Metzger \cr
Ke Xu \cr
Andrei Serafimovich \cr
}
\author{
W. S. Drysdale
}
