% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.REYN.R
\name{write.REYN}
\alias{write.REYN}
\title{Write REYN}
\usage{
write.REYN(REYN, lag_out, DirOut, analysis, tz, write_fast_data, subDir)
}
\arguments{
\item{REYN}{REYN List of objects at the end of calcualtions}

\item{lag_out}{output list from \code{wrap.lag()}}

\item{DirOut}{output directory (if subDirMonthly == TRUE, annual and monthly subfolders will be created here)}

\item{analysis}{name of analysis - used to construct file name}

\item{tz}{timezone}

\item{write_fast_data}{TRUE/FALSE should the fast (base, diff, conv, data) outputs be written to disk. They will be compressed using gzip}

\item{subDir}{one of c("none", "monthly", "daily") - default none. Should the outputs be split into monthly or daily subdirectories}
}
\description{
Writes the data associated with the REYN (classical ec) processing.
}
