% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write.REYN.R
\name{write.REYN}
\alias{write.REYN}
\title{Write REYN}
\usage{
write.REYN(REYN, lag_out, DirOut, analysis, tz, writeCompressedData, subDir)
}
\arguments{
\item{REYN}{REYN List of objects at the end of calcualtions}

\item{lag_out}{output list from \code{wrap.lag()}}

\item{DirOut}{by default is created as \code{file.path(DirWrk,"out",siteName, runID, analysis)}, can be overridden here.}

\item{analysis}{Used in file nameing}

\item{tz}{timezone, default "UTC".}

\item{writeCompressedData}{TRUE/FALSE should the larger outputs that need compression ("ACF","foot", "spec", "base", "conv", "data","diff") outputs be written to disk. They will be compressed using gzip}

\item{subDir}{one of c("none", "monthly", "daily") - default none. Should the outputs be split into monthly or daily subdirectories}
}
\description{
Writes the data associated with the REYN (classical ec) processing.
}
